Group files in an MVC folder structure
